Move date formatting logic to utils_date.py
Refactored the format_dates_in_messages function from app_kernel.py to common/utils/utils_date.py for better modularity and reuse. Updated imports to reflect the new location.
